requirements.txt
requests
Pillow
PySide6